Remove OverDrive metadata with star ratings test case. (PP-2554) (#2658)
## Description

- Removes test case for OverDrive metadata with star rating.

## Motivation and Context

Because OverDrive were scheduled to stop including the `starRating`
popularity property in their metadata by June 2025, we stopped using it
as a popularity measure with #2487 in late May 2025. The test we are
removing with this PR remained to ensure that we could process metadata
with this property until it was removed.

[Jira PP-2554]

## How Has This Been Tested?

[CI
tests](https://github.com/ThePalaceProject/circulation/actions/runs/16755340713)
pass.

## Checklist

- N/A - I have updated the documentation accordingly.
- [x] All new and existing tests passed.