MARC export updates for OCLC (PP-2618) (#2637)
## Description

Makes some of the updates requested by OCLC to MARC records produced by
MARC exporter functionality:
- 006 field (http://www.loc.gov/marc/bibliographic/bd006.html) This
field is not required but if provided, it must be coded correctly for
the electronic component. Byte 6 (Form) should be “o” for online
resource.
- 008 field (https://www.loc.gov/marc/bibliographic/bd008.html) 
- Byte 23 (Form) must be coded “o” for online resource. When this is not
coded, or coded incorrectly, the system defaults to blank which is for
print materials. This will cause the record to incorrectly match to a
print item instead of an electronic item.
- Change the following from a space to a pipe symbol (|) for “no attempt
to code”
    - §  Byte 29 (Conference publication) 
    - §  Byte 30 (Festschrift) 
    - §  Byte 31 (Inex)
    - §  Byte 33 (Literary form)
  - Byte 39 must be coded “d”.
- The [385 field](https://www.loc.gov/marc/bibliographic/bd385.html) has
invalid data in the $2 and will need to be corrected
  - Dropped field per customer guidance.
- The 385 field also has invalid data in the $2 and will need to be
corrected.
  - Dropped field per customer guidance.
- 538 fields should be deleted. According to the [Provider-neutral
Guidelines](https://www.loc.gov/aba/pcc/scs/documents/PCC-PN-guidelines.html),
for RDA, Mode of Access is not used, it is covered by Carrier type (338
field).
  - Dropped field.


## Motivation and Context

Support for OCLC vendor record contribution and additional customer
feedback (see Jira PP-2618).

## How Has This Been Tested?

- Updated tests to reflect changes.
- [CI
tests](https://github.com/ThePalaceProject/circulation/actions/runs/16572446278)
pass.

## Checklist

- N/A - I have updated the documentation accordingly.
- [x] All new and existing tests passed.