Refactor OPDS for Distributors to use OAuthOpdsRequest (PP-2723) (#2684)
## Description

Previously the OPDS for Distributors API had its own code to handle the
OPDS OAuth exchange. This PR updates it to use `OAuthOpdsRequest`
instead.

This required a change to `OAuthOpdsRequest` to handle the case where a
feed doesn't require authentication.

After this change, all of our OPDS classes are using
`BaseOpdsHttpRequest` to make requests.

## Motivation and Context

I made this change, so that it was easier to convert OPDS for
Distributors to use Celery for importing.

## How Has This Been Tested?

- New unit tests
- Tested locally

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.