Abort and cleanup if an IdentifierSet is None (PP-2721) (#2685)
## Description

Handle the case where the `mark_identifiers_unavailable` task receives
`None` instead of an `IdentifierSet`.

## Motivation and Context

In some cases one of the tasks in the `mark_identifiers_unavailable`
chord may want to abort the operation and log information, without
raising an exception.In this case we wouldn't want the
`mark_identifiers_unavailable` task to proceed, and we want to try to
clean up any data.

We handle this by returning None in the task in the chord, and handling
it appropriately within `mark_identifiers_unavailable`.

## How Has This Been Tested?

- New unit test

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.