Convert OPDS1 + OPDS for Distributors importer to Celery (PP-2721, PP-2723) (#2686)
## Description

This PR updates the OPDS1 and OPDS for Distributors importers to use
Celery.

## Motivation and Context

To make this possible, I refactored the OPDS1, OPDS for Distributors,
OPDS2, and OPDS2 + ODL importers to share more of their logic.

The goal was to strike a balance between reusing common functionality
and keeping the class hierarchy simple and easy to understand. The
result is an `OpdsExtractor` API with two implementations:

* One for OPDS1 (which also covers OPDS for Distributors)
* One for OPDS2 + ODL (which also covers OPDS2)

All OPDS variants now use a shared `OpdsImporter`.

I welcome feedback on this architecture. It took some iteration to
balance code reuse with maintainability and readability, but I think
this approach lands in a good place.

## How Has This Been Tested?

* Local testing
* New unit tests

## Checklist

* [x] Documentation has been updated
* [x] All new and existing tests pass