[PP-2812] Fix missing contributor from Boundless and Bibliotheca impo… (#2695)
…rts.

## Description
Changes made in
https://github.com/ThePalaceProject/circulation/pull/2638 had the
unintended side-effect of ignoring contributor data with sort_name but
no display name. This update fixes that bug by adding a catch block for
a ValueError when resolving the sort name and skipping if no sort name
can be resolved instead of wrongly assuming that a contributor data with
no display name is necessarily useless.

## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-2812
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Tested locally.  Unit tests updated.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [ ] All new and existing tests passed.