Refactor the apply task fixture (PP-2805) (#2702)
## Description

Refactor the Celery apply task fixture to more accurately mock the apply
task. Previously the fixture could not handle all the arguments that the
apply task could take, which was causing problems in PP-2805 when
converting the Boundless importers to the the apply task.

Note: This PR also excludes `Protocol` classes from coverage, since they
can't be called, it seems silly to mark those lines as uncovered.

## Motivation and Context

Breaking up some of the unrelated changes in the PR for PP-2805.

## How Has This Been Tested?

- Running unit tests in CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.