Retry calculate_work_presentation when work is not found (#2715)
## Description

- Retry `work.calculate_work_presentation` when `ModelNotFoundError` is
raised.
- Make `load_from_id` raise more specific `ModelNotFoundError` instead
of `PalaceValueError`
- Create a new `PalaceTypeError` exception, and make `validate_not_none`
raise `PalaceTypeError`. It doesn't seem like this should have been
raising a `ValueError` before.
- This function was only used in one place, and I made sure it wasn't
handling the exception explicitly, so it should be fine to change the
exception type.

## Motivation and Context

In my testing on GA and CA CM, we are getting a lot of this sort of
exception:
```
Task work.calculate_work_presentation[bede800f-84e6-4b65-87ba-a5b734027cd3] raised unexpected: PalaceValueError("Work with id '4598931' not found.")
```

Looking in the database after this is raised, in every case I've seen,
the work does exist. What is happening is that this task is being run
before the transaction that creates the work has been committed.
Retrying in this case should solve the issue.

## How Has This Been Tested?

- I'm planning to roll this out to CA and GA once its merged to see if
that sorts out the unhandled exceptions.

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.