Don't raise `RequestNetworkException` when retries fail (PP-2845) (#2716)
## Description

When `HTTP.request_with_timeout` retries a request, and the request runs
out of retries and still fails we return the response as usual instead
of raising a `RequestNetworkException`.

This seems like a less surprising outcome to me, and better aligns the
response in the case where we have no retries, and when we do have
retries. Previously with retries we would raise
`RequestNetworkException` and in the case of no retries we would return
the request or process it and raise a `BadResponseException`.

## Motivation and Context

In https://github.com/ThePalaceProject/circulation/pull/2709 I added
`BadResponseException` to `autoretry_for`, in order to retry when we get
a 429 status.

In testing, this didn't work because we were getting a
`RequestNetworkException`, because of the requests level retries.

This is a bit of a change in behaviour for `HTTP.request_with_timeout`,
but looking through our codebase, I don't see anywhere we are
specifically handling `RequestNetworkException` for this case, so I
don't think this will have an impact anywhere.

## How Has This Been Tested?

- Updated unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.