Increase import lock timeout to 1 hour (PP-2842) (#2718)
## Description

Increase the timeout for the import_lock to 1 hour.

## Motivation and Context

Currently this lock is using the default of 5 minutes. This is too short
for large imports, I've seen the lock expire while boundless imports are
still taking place.

I'm planning to pull in a task next sprint to speed up the boundless
imports, but I think 5 minutes is too short for this lock timeout
anyway, so I think its a good idea to bump it up.

## How Has This Been Tested?

- Will test on GA and CA CM once this is merged.

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.