Add report query definition. (PP-1888) (#2742)
## Description

Adds a mechanism for associating an id and title with a query for
tabular data for reports.

## Motivation and Context

As we add more reports, it is handy to have a consistent way of defining
and being able to lookup the queries used to produce the tabular data
that will comprise those reports.

[Jira PP-1888]

## How Has This Been Tested?

- New tests.
- Local tests pass.
- [CI
tests](https://github.com/ThePalaceProject/circulation/actions/runs/17810446785)
running.

## Checklist

- N/A - I have updated the documentation accordingly.
- [x] All new and existing tests passed.