Library collection reports support (PP-1888) (#2750)
## Description

Adds support for reports on collection data scoped to a specified
library. By default, this data is restricted to the active collections
of the library, but that can be overridden by specifying a list of any
or all collections associated with a library, whether those collections
have an active subscription or not.

## Motivation and Context

This provides support for the library title-level report requested in
PP-1888.

## How Has This Been Tested?

- New tests.
- All tests pass in locally.
- [CI
tests](https://github.com/ThePalaceProject/circulation/actions/runs/17870762495).

## Checklist

- N/A - I have updated the documentation accordingly.
- [x] All new and existing tests passed.