Drop Python 3.10 support (#2795)
## Description

This PR drops support for Python 3.10 and updates the minimum required
Python version to 3.11.

## Motivation and Context

I am doing this since we are upgrading the Python version in our docker
container to 3.12 in
https://github.com/ThePalaceProject/circulation/pull/2799.

Based on the slack discussion here:
https://lyrasis.slack.com/archives/CCY3PH8JD/p1760100023472879

We are going to try the python versioning policy:

**Support the python version we are using in the repos docker container,
and any more recent versions. Attempt to add the most recent version as
soon as our dependencies allow it.**

So after updating the docker image to use Python 3.12, we can drop 3.10
and 3.11 support. I'm going to do this in two separate PRs. This one
will drop Python 3.10 support and I'll handle 3.11 support in a follow
up.

This allows us to:
- Remove compatibility shims (backports.strenum, typing_extensions)
- Use native Python 3.11+ features (StrEnum, Self type)
- Simplify the codebase by removing version checks

## How Has This Been Tested?

- Pre-commit hooks pass on all files
- Configuration changes have been validated
- CI workflows updated to run on Python 3.11 and 3.12

## Checklist

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.